use Integer instead of Int64
Some length() are returning Int32. Also,
CI_HS1 do not use n_excitedDOF

+++ src/conjugated.jl
@@ -2,8 +2,8 @@
 # complex conjugate and the initial conditions are not homogeneous
 function CbFFconj!(response::Array{TC}, K::AbstractArray, M::AbstractArray, 
                    CbF::AbstractArray, F11::AbstractArray, F11_squared::AbstractArray,
-                   expF11_delta::AbstractArray, dim::Int64,
-                   n_excitedDOF::Int64, times::Ts, n_times::Int64, input_vectors::Matrix{TF}, 
+                   expF11_delta::AbstractArray, dim::Integer,
+                   n_excitedDOF::Integer, times::Ts, n_times::Integer, input_vectors::Matrix{TF}, 
                    loads::Matrix{TF},int_loads::Matrix{TF},C2::Vector) where {Ts,TC,TF}
  
      # Calculates matrix Omega_2
